requirements.txt
python-telegram-bot[job-queue]
pycoingecko
python-decouple
aiohttp
pandas